AddServices

+++ UsedCarHub.BusinessLogic/DTOs/CarDto.cs
@@ -10,7 +10,7 @@ namespace UsedCarHub.BusinessLogic.DTOs
         public string Model { get; set; }
         public int YearOfProduction { get; set; }
         public Transmission TransmissionType { get; set; }
-        public int EngineCapacity { get; set; }
+        public float EngineCapacity { get; set; }
         public int Mileage { get; set; }
     }
 }
